+++ packages/frontend/src/renderer/containers/widgets/loadingPanel/loadingMessages.ts
@@ -1,0 +1,6 @@
+export enum LoadingMessages {
+  CreateCommunity = 'Creating community',
+  JoinCommunity = 'Joining community',
+  StartApp = 'Starting Quiet',
+  RetryRegistration = 'Retrying registration'
+}