+++ packages/nectar/src/sagas/publicChannels/subscribeToAllTopics/subscribeToAllTopics.saga.test.ts
@@ -43,7 +43,7 @@ describe('subscribeToAllTopicsSaga', () => {
     hiddenService: { onionAddress: 'onionAddress', privateKey: 'privateKey' },
     dmKeys: { publicKey: 'publicKey', privateKey: 'privateKey' },
     peerId: { id: 'peerId', pubKey: 'pubKey', privKey: 'privKey' },
-    zbayNickname: '',
+    nickname: '',
     userCsr: undefined,
     userCertificate: ''
   }