+++ packages/nectar/src/sagas/users/users.adapter.ts
@@ -1,5 +1,5 @@
 import { createEntityAdapter } from '@reduxjs/toolkit'
-import { keyFromCertificate } from '@zbayapp/identity'
+import { keyFromCertificate } from '@quiet/identity'
 import { Certificate } from 'pkijs'
 
 export const certificatesAdapter = createEntityAdapter<Certificate>({