+++ packages/waggle/src/storage/testUtils/storageBasic.ts
@@ -19,12 +19,12 @@ export class StorageBasic extends Storage {
   public declare options: StorageTestSnapshotOptions
 
   constructor(
-    zbayDir: string,
+    quietDir: string,
     io: any,
     communityId: string,
     options?: Partial<StorageTestSnapshotOptions>
   ) {
-    super(zbayDir, io, communityId, options)
+    super(quietDir, io, communityId, options)
     this.options = {
       ...new StorageTestSnapshotOptions(),
       ...options